added validation

+++ src/tasks/tasks.resolver.ts
@@ -1,4 +1,9 @@
 import {
+  BadRequestException,
+  NotFoundException,
+  ValidationPipe,
+} from '@nestjs/common';
+import {
   Args,
   Mutation,
   Parent,
@@ -7,7 +12,13 @@ import {
   Resolver,
 } from '@nestjs/graphql';
 import { Prisma, Task, User } from '@prisma/client';
-import { Nullable } from 'src/types/typescript.types';
+import {
+  CreateTaskInput,
+  TaskWhereUniqueInput,
+  UpdateTaskInput,
+} from 'src/typescript/gql-generated-types';
+
+import { Nullable } from 'src/typescript/types';
 import { UsersService } from 'src/users/users.service';
 
 import { TasksService } from './tasks.service';
@@ -26,7 +37,7 @@ export class TasksResolver {
 
   @Query('task')
   async getTask(
-    @Args('input') input: Prisma.TaskWhereUniqueInput,
+    @Args('input') input: TaskWhereUniqueInput,
   ): Promise<Nullable<Task>> {
     return this.tasksService.getTask(input);
   }
@@ -34,30 +45,51 @@ export class TasksResolver {
   @ResolveField('user')
   async getUser(@Parent() task: Task): Promise<User> {
     const { userId: id } = task;
-
     return this.usersService.getUser({ id });
   }
 
   @Mutation()
-  async createTask(
-    @Args('input') input: Prisma.TaskCreateInput,
-  ): Promise<Task> {
+  async createTask(@Args('input') input: CreateTaskInput): Promise<Task> {
+    const { userId: id } = input;
+
+    const user = await this.usersService.getUser({ id });
+
+    if (!user) {
+      throw new BadRequestException('Provided user does not exist!');
+    }
+
     return this.tasksService.createTask(input);
   }
 
   @Mutation()
   async updateTask(
     @Args('input')
-    input: Prisma.TaskUpdateInput,
+    input: UpdateTaskInput,
   ): Promise<Nullable<Task>> {
+    const { id } = input;
+
+    const task = await this.tasksService.getTask({ id });
+
+    if (!task) {
+      throw new BadRequestException('Task does not exist!');
+    }
+
     return this.tasksService.updateTask(input);
   }
 
   @Mutation()
   async deleteTask(
     @Args('input')
-    input: Prisma.TaskWhereUniqueInput,
+    input: TaskWhereUniqueInput,
   ): Promise<Nullable<Task>> {
+    const { id } = input;
+
+    const task = await this.tasksService.getTask({ id });
+
+    if (!task) {
+      throw new BadRequestException('Task does not exist!');
+    }
+
     return this.tasksService.deleteTask(input);
   }
 }
