+++ src/users/users.resolver.ts
@@ -1,6 +1,10 @@
+import { NotFoundException } from '@nestjs/common';
 import { Resolver, Args, Mutation, Query } from '@nestjs/graphql';
 import { Prisma, User } from '@prisma/client';
-import { Nullable } from 'src/types/typescript.types';
+import { UserWhereUniqueInput } from 'src/typescript/gql-generated-types';
+import { Nullable } from 'src/typescript/types';
+import { CreateUserDto } from './dtos/create-user.dto';
+import { UpdateUserDto } from './dtos/update-user.dto';
 import { UsersService } from './users.service';
 
 @Resolver()
@@ -14,31 +18,45 @@ export class UsersResolver {
 
   @Query('user')
   async getUser(
-    @Args('input') input: Prisma.UserWhereUniqueInput,
+    @Args('input') input: UserWhereUniqueInput,
   ): Promise<Nullable<User>> {
     return this.usersService.getUser(input);
   }
 
   @Mutation()
-  async createUser(
-    @Args('input') input: Prisma.UserCreateInput,
-  ): Promise<User> {
+  async createUser(@Args('input') input: CreateUserDto): Promise<User> {
     return this.usersService.createUser(input);
   }
 
   @Mutation()
   async updateUser(
     @Args('input')
-    input: Prisma.UserUpdateInput,
+    input: UpdateUserDto,
   ): Promise<Nullable<User>> {
+    const { id } = input;
+
+    const user = await this.usersService.getUser({ id });
+
+    if (!user) {
+      throw new NotFoundException('User does not exist.');
+    }
+
     return this.usersService.updateUser(input);
   }
 
   @Mutation()
   async deleteUser(
     @Args('input')
-    input: Prisma.UserWhereUniqueInput,
+    input: UserWhereUniqueInput,
   ): Promise<Nullable<User>> {
+    const { id } = input;
+
+    const user = await this.usersService.getUser({ id });
+
+    if (!user) {
+      throw new NotFoundException('User does not exist.');
+    }
+
     return this.usersService.deleteUser(input);
   }
 }
